docs:delete star.md
