--- Stylish-Front-End-main/src/components/Stars/star.md
@@ -1,7 +1,0 @@
-import Stars from "./components/Stars";
-
-<Star size={40} rate={4} space={8} />
-
-size 輸入幾 px
-rate 輸入要幾個星星(不能超過 5，也不能小於 0)
-space 星星之間的空格大小
